refactor: 👷 helper function in /lib/utils/
